requirements.txt
streamlit
SpeechRecognition
gTTS
python-dotenv
openai
pyaudio ; platform_system == "Windows"
